refactor: move out of x/swingset/internal
