+++ packages/cosmic-swingset/proto/agoric/swingset/genesis.proto
@@ -3,7 +3,7 @@ package agoric.swingset;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/Agoric/cosmic-swingset/x/swingset/internal/types";
+option go_package = "github.com/Agoric/cosmic-swingset/x/swingset/types";
 
 message GenesisState {
     option (gogoproto.equal) = false;
