+++ packages/cosmic-swingset/proto/agoric/swingset/msgs.proto
@@ -4,7 +4,7 @@ package agoric.swingset;
 import "gogoproto/gogo.proto";
 import "ibc/core/channel/v1/channel.proto";
 
-option go_package = "github.com/Agoric/cosmic-swingset/x/swingset/internal/types";
+option go_package = "github.com/Agoric/cosmic-swingset/x/swingset/types";
 
 service Msg {
   rpc DeliverInbound(MsgDeliverInbound) returns (MsgDeliverInboundResponse);
