+++ packages/cosmic-swingset/proto/agoric/swingset/query.proto
@@ -6,7 +6,7 @@ import "agoric/swingset/storage.proto";
 import "cosmos/base/query/v1beta1/pagination.proto";
 import "google/api/annotations.proto";
 
-option go_package = "github.com/Agoric/cosmic-swingset/x/swingset/internal/types";
+option go_package = "github.com/Agoric/cosmic-swingset/x/swingset/types";
 
 // Query provides defines the gRPC querier service
 service Query {
