+++ packages/cosmic-swingset/x/swingset/genesis.go
@@ -3,7 +3,7 @@ package swingset
 import (
 	// "fmt"
 
-	"github.com/Agoric/cosmic-swingset/x/swingset/internal/types"
+	"github.com/Agoric/cosmic-swingset/x/swingset/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	abci "github.com/tendermint/tendermint/abci/types"
 )
