+++ packages/cosmic-swingset/x/swingset/module.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/Agoric/cosmic-swingset/x/swingset/client/cli"
 	"github.com/Agoric/cosmic-swingset/x/swingset/client/rest"
-	"github.com/Agoric/cosmic-swingset/x/swingset/internal/keeper"
-	"github.com/Agoric/cosmic-swingset/x/swingset/internal/types"
+	"github.com/Agoric/cosmic-swingset/x/swingset/keeper"
+	"github.com/Agoric/cosmic-swingset/x/swingset/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
